retrodemo: Seperate into modules
Makes the code a bit easier to digest.
